+++ src/presentation/controllers/login/login-protocols.ts
@@ -1,4 +1,4 @@
 export * from '../../../domain/usecases/authentication'
 export * from '../../protocols'
 export * from '../../protocols/email-validator'
-export * from '../../helpers/validators/validation'
+export * from '../../protocols/validation'
